Move App and setup initial heading
