+++ src/App/App.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import './App.css';
+
+function App() {
+  return (
+    <div className="App">
+      <header className="App-header">
+        <h1>Time My Life</h1>
+      </header>
+    </div>
+  );
+}
+
+export default App;
